refactor(media): move MediaCreatedEvent to media bounded context
MediaCreatedEvent is specific to the media module, not a generic
building block. Move it to src/modules/media/domain/events.py and
update all imports. DomainEvent base class stays in building_blocks.
